CEDS-1246 Remove traits for mapping strategy
