+++ test/services/WcoMetadataJavaMappingStrategySpec.scala
@@ -26,7 +26,7 @@ class WcoMetadataJavaMappingStrategySpec extends WordSpec with Matchers with Sch
 
   "WcoMetadataJavaMappingSpec" should {
     "produce metadata" in {
-      val mapper = new WcoMetadataMapper with WcoMetadataJavaMappingStrategy
+      val mapper = new WcoMetadataMapper
       val metaData = mapper.produceMetaData(SupplementaryDeclarationDataSpec.cacheMapAllRecords, supplementaryChoice)
 
       metaData.getWCOTypeName.getValue shouldBe SchemaMandatoryValues.wcoTypeName
@@ -38,26 +38,26 @@ class WcoMetadataJavaMappingStrategySpec extends WordSpec with Matchers with Sch
     }
 
     "retrieve a DUCR based on the produced metadata" in {
-      val mapper = new WcoMetadataMapper with WcoMetadataJavaMappingStrategy
+      val mapper = new WcoMetadataMapper
       val metaData = mapper.produceMetaData(SupplementaryDeclarationDataSpec.cacheMapAllRecords, supplementaryChoice)
 
       mapper.declarationUcr(metaData) should be(Some("8GB123456789012-1234567890QWERTYUIO"))
     }
 
     "retrieve a LRN based on the produced metadata" in {
-      val mapper = new WcoMetadataMapper with WcoMetadataJavaMappingStrategy
+      val mapper = new WcoMetadataMapper
       val metaData = mapper.produceMetaData(SupplementaryDeclarationDataSpec.cacheMapAllRecords, supplementaryChoice)
 
       mapper.declarationLrn(metaData) should be(Some("123LRN"))
     }
 
     "marshall the metadata correctly" in {
-      val mapper = new WcoMetadataMapper with WcoMetadataJavaMappingStrategy
+      val mapper = new WcoMetadataMapper
       val metaData = mapper.produceMetaData(SupplementaryDeclarationDataSpec.cacheMapAllRecords, supplementaryChoice)
 
       mapper.toXml(metaData) should include(Source.fromURL(getClass.getResource("/wco_dec_metadata.xml")).mkString)
 
-      validateXmlAgainstSchema(mapper.serialise(metaData))
+      validateXmlAgainstSchema(mapper.toXml(metaData))
     }
   }
 }
